feat: Initial Commit
Started to work on the augmented system. All augmented projects will be part of my commercial projects. The first one being the shop system which we are also going to use inside the hub.

Currently I implemented a database connection and the temporary shop fetching, which both seem to work perfectly.